% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femeta.R
\name{femeta.array}
\alias{femeta.array}
\title{Array Method of Improved Inference for Fixed-effects Meta-analysis}
\usage{
\method{femeta}{array}(x, correction_factor = 0.01, drop00 = FALSE,
  estimator = "log-CMH", vtype = "Het", alternative = "two.sided",
  conf.level = 0.95, null.value = 0, ...)
}
\arguments{
\item{x}{vector to specify the 2x2 table frequencies (upper left cell).}

\item{correction_factor}{A scalar used for zero-cell correction. Intuitively it is the total number of
subjects added to each group of a study. A small correction factor induces less bias. The default is 0.01.
See 'Details'.}

\item{drop00}{logical indicating whether studies with no cases/events (or only cases) in both groups
should be dropped when calculating the observed outcomes of the individual studies. See 'Details'.}

\item{estimator}{a character string indicating which estimator is used to summarize to overall association (either "log-CMH",
"Woolf" or "CEMLE"). See 'Details'.}

\item{vtype}{a character string indicating whether to assume homogeneity when calculating the variance (either "Het" or "Hom).}

\item{alternative}{a character string indicating the alternative hypothesis and must be one of "two.sided", "greater"
or "less". You can specify just the initial letter.}

\item{conf.level}{a scalar indicating the confidence level for the returned confidence interval}

\item{null.value}{scalar indicating the value of the parameter of interest (some summary of the true effect sizes)}

\item{...}{other arguments passed to \link{femeta.default}.}
}
\description{
Array Method of Improved Inference for Fixed-effects Meta-analysis
}
\examples{
X <- array(sample(1:20, size = 12, replace = TRUE), dim = c(2, 2, 3))
femeta(X)
}
