% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femeta.R
\name{femeta.default}
\alias{femeta.default}
\title{Default Method of Improved Inference for Fixed-effects meta-analysis}
\usage{
\method{femeta}{default}(ai, bi, ci, di, n1i, n2i, x1i, x2i, data, slab,
  subset, correction_factor = 0.01, drop00 = FALSE,
  estimator = "log-CMH", vtype = "Het", alternative = "two.sided",
  conf.level = 0.95, null.value = 0)
}
\arguments{
\item{ai}{vector to specify the 2x2 table frequencies (upper left cell).}

\item{bi}{vector to specify the 2x2 table frequencies (upper right cell).}

\item{ci}{vector to specify the 2x2 table frequencies (lower left cell).}

\item{di}{vector to specify the 2x2 table frequencies (lower right cell).}

\item{n1i}{vector to specify the group sizes or column totals (first group/column).}

\item{n2i}{vector to specify the groups sizes or column totals (second group/column).}

\item{x1i}{vector to specify the number of events (first group).}

\item{x2i}{vector to specify the number of evetns (second group).}

\item{data}{optional data frame containing the variables given to the arguments above.}

\item{slab}{optional vector with labels for the studies.}

\item{subset}{optional vector indicating the subset of studies that should be used.
This can be a logical vector or a numeric vector indicating the indices of the studies to include.}

\item{correction_factor}{A scalar used for zero-cell correction. Intuitively it is the total number of
subjects added to each group of a study. A small correction factor induces less bias. The default is 0.01.
See 'Details'.}

\item{drop00}{logical indicating whether studies with no cases/events (or only cases) in both groups
should be dropped when calculating the observed outcomes of the individual studies. See 'Details'.}

\item{estimator}{a character string indicating which estimator is used to summarize to overall association (either "log-CMH",
"Woolf" or "CEMLE"). See 'Details'.}

\item{vtype}{a character string indicating whether to assume homogeneity when calculating the variance (either "Het" or "Hom).}

\item{alternative}{a character string indicating the alternative hypothesis and must be one of "two.sided", "greater"
or "less". You can specify just the initial letter.}

\item{conf.level}{a scalar indicating the confidence level for the returned confidence interval.of the}

\item{null.value}{scalar indicating the value of the parameter of interest (some summary of the true effect sizes)}
}
\description{
Default Method of Improved Inference for Fixed-effects meta-analysis
}
