% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femeta.R
\name{print.femeta}
\alias{print.femeta}
\title{Printing results of fixed-effects meta-analysis}
\usage{
\method{print}{femeta}(x, digits = getOption("digits"), prefix = "\\t")
}
\arguments{
\item{x}{object of class "femeta"}

\item{digits}{number of significant digits to be used}

\item{prefix}{string, passed to strwrap for displaying the method component of the \code{femeta} object.}
}
\description{
Printing results of fixed-effects meta-analysis
}
